feat: add /health endpoint (#8)
This pull request includes updates to the Go linter version in the
GitHub Actions workflow and adds a new health check endpoint to the
backend server.

### Workflow updates:

*
[`.github/workflows/lint.yml`](diffhunk://#diff-107e910e9f2ebfb9a741fa10b2aa7100cc1fc4f5f3aca2dfe78b905cbd73c0d2L47-R47):
Updated the Go linter (`golangci-lint`) version from `v1.61.0` to
`v1.63.4` to ensure compatibility with the latest features and fixes.

### Backend server enhancements:

*
[`backend/server.go`](diffhunk://#diff-ff9229da69d0693af25e0f7bfb82853d7eb82d68d0f1a9938c866232acb32debR14-R16):
Added a new `/health` endpoint to the Echo server, which returns an "OK"
response for health checks.
*
[`backend/server.go`](diffhunk://#diff-ff9229da69d0693af25e0f7bfb82853d7eb82d68d0f1a9938c866232acb32debR5):
Added an empty line to organize imports for better readability.